fix: ensure username is always initialized correctly
There was actually no code (in the current revision) that ever persisted
the user name. It was always presumed to already be on disk.

Presumably this worked in the past and broke after some refactoring.

The exact error looked something like this:

Uncaught Error: Typed property Grandeljay\Availability\UserAvailability::$userName must not be accessed before initialization in src/Grandeljay/Availability/UserAvailability.php:165