feat: add `--install` startup argument
Usually when starting the bot I get rate limited due to the commands being all removed and re-added again. The official discord API docs states that the code for adding the commands should be removed after adding them.

Now, we can simple run `php availability.php --install` if the added or removed a command and need it to update in discord.